+++ starkinfra/issuinginvoice/log/__log.py
@@ -18,11 +18,11 @@ class Log(Resource):
     - created [datetime.datetime]: creation datetime for the log. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
     """
 
-    def __init__(self, id, created, type, invoice):
+    def __init__(self, id, invoice, type, created):
         Resource.__init__(self, id=id)
 
-        self.type = type
         self.invoice = from_api_json(_issuing_invoice_resource, invoice)
+        self.type = type
         self.created = check_datetime(created)
 
 
@@ -47,7 +47,7 @@ def query(ids=None, limit=None, after=None, before=None, types=None, user=None):
     Receive a generator of issuinginvoice.Log objects previously created in the Stark Infra API
     ## Parameters (optional):
     - ids [list of strings, default None]: list of IssuingInvoice ids to filter logs. ex: ["5656565656565656", "4545454545454545"]
-    - limit [integer, default 100]: maximum number of objects to be retrieved. Unlimited if None. ex: 35
+    - limit [integer, default None]: maximum number of objects to be retrieved. Unlimited if None. ex: 35
     - after [datetime.date or string, default None] date filter for objects created only after specified date. ex: datetime.date(2020, 3, 10)
     - before [datetime.date or string, default None] date filter for objects created only before specified date. ex: datetime.date(2020, 3, 10)
     - types [list of strings, default None]: filter for log event types. ex: ["created", "credited", "expired", "overdue", "paid"]